restructure commitments pallet
